FCBHDBP-577 Updated the endpoint to translate a plan to include the playlist item duration value.
