+++ composer.json
@@ -15,26 +15,26 @@
         "algolia/algoliasearch-client-php": "3.3.2",
         "appstract/laravel-opcache": "4.0.2",
         "awobaz/compoships": "^2.2",
-        "aws/aws-sdk-php": "3.269.2",
+        "aws/aws-sdk-php": "3.272.0",
         "biscolab/laravel-recaptcha": "6.0.0",
         "bugsnag/bugsnag-laravel": "2.26.0",
         "caneara/quest": "3.0.0",
         "curl/curl": "2.5.0",
-        "doctrine/dbal": "3.6.2",
+        "doctrine/dbal": "3.6.3",
         "geoip2/geoip2": "2.13.0",
         "google/recaptcha": "1.3.0",
-        "guzzlehttp/guzzle": "7.5.1",
+        "guzzlehttp/guzzle": "7.7.0",
         "intervention/image": "2.7.2",
         "josiasmontag/laravel-localization": "1.0.1",
-        "laravel/framework": "9.52.7",
+        "laravel/framework": "9.52.9",
         "laravel/slack-notification-channel": "2.5.0",
-        "laravel/socialite": "5.6.1",
+        "laravel/socialite": "5.6.3",
         "laravel/tinker": "2.8.1",
         "maennchen/zipstream-php": "2.4.0",
         "sentry/sentry-laravel": "2.14.2",
         "socialiteproviders/generators": "dev-php-8",
         "socialiteproviders/google": "4.1.0",
-        "spatie/array-to-xml": "3.1.5",
+        "spatie/array-to-xml": "3.1.6",
         "spatie/eloquent-sortable": "4.0.1",
         "spatie/laravel-fractal": "6.0.3",
         "sunra/php-simple-html-dom-parser": "1.5.2",
@@ -48,13 +48,13 @@
     "require-dev": {
         "barryvdh/laravel-debugbar": "3.8.1",
         "barryvdh/laravel-ide-helper": "2.13.0",
-        "fakerphp/faker": "1.21.0",
+        "fakerphp/faker": "1.22.0",
         "filp/whoops": "2.15.2",
         "friendsofphp/php-cs-fixer": "3.13.0",
         "itsgoingd/clockwork": "5.1.12",
         "laravel/legacy-factories": "1.3.2",
-        "mockery/mockery": "1.5.1",
-        "phpunit/phpunit": "9.6.7",
+        "mockery/mockery": "1.6.2",
+        "phpunit/phpunit": "9.6.8",
         "spatie/laravel-tail": "4.4.2",
         "squizlabs/php_codesniffer": "3.7.2"
     },
